% $Id$

\name{tclish-package}
\alias{tclish-package}
\docType{package}
\title{Tcl/Tk Additions}
\description{
  A package containing a variety of Tcl/Tk methods used for GUI development.
}
\details{
  \tabular{ll}{
    Package: \tab tclish\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("tclish")$Version}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("tclish")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("tclish")$License}\cr
  }

  For a complete list of functions, use \code{library(help="tclish")}.\cr
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="tclish"))}.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

