\name{tclish-package}
\alias{tclish-package}
\docType{package}
\title{Tcl/Tk Additions}
\description{
  A package containing a variety of Tcl/Tk methods used for GUI development.
}
\details{
  \tabular{ll}{
    Package: \tab tclish\cr
    Type: \tab Package\cr
    Version: \tab 0.8.3\cr
    Date: \tab 2011-11-20\cr
    License: \tab See LICENSE for license information\cr
  }

  For a complete list of functions, use \code{library(help="tclish")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="tclish"))}.
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{package}

